Correct more linters
